// frontend/src/api/models/Metadata.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Metadata = {
    id?: string;
    filename: string;
    full_path: string;
    create_date: string;
    keywords: Array<string>;
    description: string;
    title: string;
};